+++ lib/glimmer-dsl-libui/ext/glimmer.rb
@@ -1,0 +1,7 @@
+module Glimmer
+  class << self
+    def included(klass)
+      klass.extend(Glimmer)
+    end
+  end
+end